INTEGER to BIGINT
Alterado o DataTypes dos atributos "id" que fazem referencia a "id" de outras tabelas de INTEGER pra BIGINT

+++ src/models/tribut_cofins.ts
@@ -27,7 +27,7 @@ TributCofins.init({
    primaryKey: true,
  },
  id_tribut_configura_of_gt: {
-   type: DataTypes.INTEGER,
+   type: DataTypes.BIGINT,
    references: { model: "tribut_configura_of_gt", key: "id" }
  },
  cst_cofins: {
